33 agent 백엔드 에이전트응답 db저장

+++ franchise/backend/src/main/java/com/franchise/backend/briefing/dto/Event48hDto.java
@@ -12,10 +12,10 @@ import lombok.*;
 public class Event48hDto {
 
     @JsonProperty("event_id")
-    private Integer eventId;
+    private Long eventId;
 
     @JsonProperty("store_id")
-    private Integer storeId;
+    private Long storeId;
 
     @JsonProperty("store_name")
     private String storeName;
